Add author information and links
This commit modifies the index page to reflect the multiple authors now
working on the project. As e-m-ilia was the original creator of the
index page, her information comes first, followed by the others in the
order that I read their usernames from github.

The Github link in the footer now links to the repo, which is private,
so I gave it the (Login required) disclaimer as with the duo link.