Update index links, tweak carousel
The article added by jcgit786 was intended to be linked from the tiles
in the middle of the site, rather than the carousel. The extra carousel
slide has been left in, tweaked slightly to be more consistent with the
other slides, but I'm thinking about removing it entirely to meet the
original concept.